fix: `tickCount` in tick sampling
