+++ packages/vutils-extension/src/transform/tick-data/index.ts
@@ -2,7 +2,7 @@ import type { BandScale, ContinuousScale, IBaseScale } from '@visactor/vscale';
 // eslint-disable-next-line no-duplicate-imports
 import { isContinuous, isDiscrete } from '@visactor/vscale';
 import { continuousTicks } from './continuous';
-import { linearDiscreteTicks } from './discrete/linear-quick';
+import { linearDiscreteTicks } from './discrete/linear';
 import { polarAngleAxisDiscreteTicks } from './discrete/polar-angle';
 import type { ICartesianTickDataOpt, IPolarTickDataOpt, ITickData, ITickDataOpt } from './interface';
 import { convertDomainToTickData } from './util';
